requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# Data Validation
pydantic==2.10.3
pydantic[email]==2.10.3

# RAG & AI - Compatible versions
langchain==0.3.13
langchain-community==0.3.13
langchain-text-splitters==0.3.4
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
pydantic-settings==2.6.1
